# internal/api/openapi_types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package api

import (
	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// ErrorResponse Standard error response.
type ErrorResponse struct {
	Detail *string `json:"detail,omitempty"`
	Status int     `json:"status"`
	Title  string  `json:"title"`
}

// UserResponse A user profile. Sourced from JWT claims (for /me) or Keycloak Admin API (for /users).
type UserResponse struct {
	Email *openapi_types.Email `json:"email,omitempty"`

	// Enabled Whether the user account is active in Keycloak.
	Enabled   bool    `json:"enabled"`
	FirstName *string `json:"firstName,omitempty"`

	// Id Keycloak user ID.
	Id       openapi_types.UUID `json:"id"`
	LastName *string            `json:"lastName,omitempty"`

	// Organizations Keycloak organization aliases the user belongs to.
	Organizations *[]string `json:"organizations,omitempty"`
	Username      string    `json:"username"`
}

// AdminNotConfigured Standard error response.
type AdminNotConfigured = ErrorResponse

// Forbidden Standard error response.
type Forbidden = ErrorResponse

// Unauthorized Standard error response.
type Unauthorized = ErrorResponse

// ListUsersParams defines parameters for ListUsers.
type ListUsersParams struct {
	// Organization Filter users by Keycloak organization alias.
	Organization *string `form:"organization,omitempty" json:"organization,omitempty"`
}